Extract ItemPlayButton
